+++ @navikt/core/react/src/form/Fieldset/index.ts
@@ -1,5 +1,2 @@
-export {
-  default as Fieldset,
-  type FieldsetProps,
-  FieldsetContext,
-} from "./Fieldset";
+export { default as Fieldset, type FieldsetProps } from "./Fieldset";
+export { FieldsetContext } from "./context";
