+++ @navikt/core/react/src/form/useFormField.ts
@@ -1,7 +1,7 @@
 import cl from "clsx";
 import React, { useContext } from "react";
 import { useId } from "../util";
-import { FieldsetContext } from "./Fieldset";
+import { FieldsetContext } from "./Fieldset/context";
 
 export interface FormFieldProps {
   /**
